pyproject.toml
[project]
name = "weather-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.110.0",
    "pydantic-settings>=2.2.0",
    "pydantic>=2.7.0",
    "uvicorn[standard]>=0.29.0",
    "httpx==0.23.3",
    "jinja2>=3.1.3",
    "python-dotenv>=1.0.1",
    "pytest>=8.1.0",
    "pytest-asyncio>=0.23.5",
    "youtube-search-python>=1.6.6",
    "python-amazon-paapi>=3.2.0",
]
